requirements.txt
requests
python-dotenv
functions-framework==3.*